Do not keep redundant RSS feeds
A snapshot is identified by hashing the GUID hashes of all items in an
RSS feed. A naive hash of the entire RSS file does not work as some RSS
feeds include unique content per request even though there is no
material change to the feed.

This has left the feed.ts file a bit messy... lots of ambiguous ways of
what a "snapshot" is and what functions produce channels from them.

Issue #3