Ensure singleton access to the archive
A lock file used at the archive level to ensure that only one process at
a time can access the archive.

This is necessary to allow podstash to run at frequent intervals.
Without this it is possible that multiple instances may attempt to
download the same episodes concurrently leading to corruption.

Issue #8